refs #308
Further work towards better support of the development workflow of TG-based applications in the area of main menu construction. 

This change includes decoupling of the main menu related API. Among different changes this includes introduction of the '''getUser''' method as part of the  IEntityDao contract and separation of the IMainMenuItemController into two independent interfaces IMainMenuItemController and IMainMenuStructureBuilder.

The new interface IMainMenuStructureBuilder has been provided with both remote and local implementations offering support for deployment and development workflows.

git-svn-id: http://kid/svn/trident-genesis/trunk@3353 50ee8237-08a1-5d4c-b143-727193748607