refs #397
Reinstated the validation code for one-to-many association as part of entity instantiation, but in modified form -- only persisted entity types are required to maintain valid one-to-many associations. 

Synthetic entity types do not have that requirement. For example, a synthetic entity that is used to represent some unit of work, saving which a bunch if persisted entities should be updated.



git-svn-id: http://kid/svn/trident-genesis/trunk@4111 50ee8237-08a1-5d4c-b143-727193748607