Introduced annotation WhenNullMessage.
This annotation should be used to annotate either an entity object or its properties where some custom message should be provided when either key or the annotated property has value null (i.e. has no value).

Such custom message can then be used as required whenever necessary...
For example, it is used as part of property editors to display a custom message when property is null.

git-svn-id: http://kid/svn/trident-genesis/trunk@4953 50ee8237-08a1-5d4c-b143-727193748607