#1554 Remove 'rejectionhandled' event handler.
'rejectionhandled' event for promise listening is of no use for us because it is fired only when promise didn't have .catch() handler and was already fulfilled (i.e. rejected) but gets it .catch() handler later in the code. This situation leads to 'unhadledrejection' event which is not desirable for us. So .catch() clauses have to be chained during promise creation and that is the case for all our promises.