#777 Enhanced GraphQLService to support @DenyIntrospection.
As a note, types that have @DenyIntrospection still needs to be present in the GraphQL dictionary. This actually makes them visible in introspection queries. However, the implemented change limits this expose to the type's title and description, while not exposing its properties.

Another note worth making, is that properties of an entity type with introspection denied, are not exposed, even though an entity where they are declared may not have introspection denied.