#1254 Introducedd a redirect for unauthorised GET requests to the login resource.
This change work for both RSO and SSO authentication modes due to the most recent chagnes to the login resource. The main objective of this change is to minimise user frustration when tyring to open some link to a TG-based application that does not include /login and there is no active user session.