#849 Further refactoring to simplife and harden processing.
The major change in this commit is the removal of the analysis of dependencies between meta-models to identify those that need to be re-generated upon activation/deaction of existing meta-models. However, in practice it would appear that such regeneration takes place automatically due to additional rounds of annotation processing, instigated by Eclipse.