#1855 Refactored EntityElement and MetaModelElement.
These classes used to be implemented as simple wrappers composed of a
TypeElement. These changes introduce a new abstract class that serves as
a parent for platform-defined Element implementations. This class,
ForwardingElement, can be parameterized with a subtype of Element. It
serves the role of a forwarder. ForwardingTypeElement, for example, will
forward all calls to TypeElement. This new type hierarchy allows us to
provide our own implementations of Element while preserving
polymorphism.