#1692 Adapted type renaming with ByteBuddy.
First commit in a series of migrating from ASM to ByteBuddy in the
implementation of TypeMaker API. TypeMaker is parameterized with the
original type. DynamicEntityClassLoader is no longer used except as an
entry point (will adjust later), thus no more caching of byte[].
However, caching is important, so will try to adjust later.