#1692 Refactored DynamicEntityClassLoader to work with ByteBuddy.
In order to maintain a cache of dynamically generated types AND use a
single class loader for all of them a custom class loader (compatible
with ByteBuddy) is needed. Although caching is supported by ByteBuddy
itself, it leads to the creation of a separate class loader for each
dynamically created type, which is less efficient.