#1692 Refactoring NewProperty. Deprecating members.
Previous implementation of NewProperty was designed to work closely with
ASM. It relied on the control field `changeSignature` to determine
whether the raw type or the type argument of the property should be
modified. It also lacked information about the actual type arguments.
The usage of `changeType` and `changeTypeSignature` was confusing, since
it didn't actually retain any information from the existing property.
These methods will be replaced by a convenient alternative.