#1882 Support for the TG Test App to run againts alternative databases.
It is now possible to run PopulateDb and Start against either PostgreSQL or SQL Server.
It is required to provide system properties 'databseUri', 'databaseUser' and 'databasePasswd', which are not stored in the default files for application properties.
There are now 2 default application properties files -- one for PostgreSQL and one for SQL Server -- which are selected automatically, based on the databaseUri (i.e., similar as for tests).
It is possible to provide an different application properties file, if required. If such file contains values for 'hibernate.connection.url', 'hibernate.connection.username' and 'hibernate.connection.password', then the file values will override those provided as system properties.