#1692 Enhanced EntityUtils to correctly support the new way of entity generation (EQL implications).
It became required to traverse the type hierarchy instead of relying on declared fields when identifying whether an entity type represents Synthetic Entity or not.
For the same reason identification of Synthetic Entities based on Persistent requires now hierarchy traversal to correctly identify the persistent nature of generated types, which are based on an entity that is Synthetic with Persisten nature.