#1937 Refactore GraphQLService to use IAuthorisationModel without a provider.
This became possible due to the thread-safety and a singleton nature of ServerAuthorisationModel. One might argue that we should not depend on such implementation details. However, the introduction of a Provider was also due to an implementation detail of GraphQLService being instantiated in a singleton scope. Also, there are other dependencies of the same nature – ICompanionObjectFinder and ISecurityTokenProvider must be thread-safe.