#849 Adjusted naming for fields that represent entity props in meta-models.
This change is required to avoid name collision between generated fields that represent props from domain entities, and fields that are part of the core meta-data structure.
The most recent example of this was property 'path', which collided with meta-model field of the same name, which represents dot-notated paths in the entity graph. As the functionality of meta-modelling becomes richer, new similar collisions could be possible. This change should prevent such name collisions.