#2178 Employ a general approach of associating metadata (e.g., labels) with grammar terms
This approach uses an external storage for metadata that is associated
with a particular BNF instance. Unlike with the previous approach,
original Term instances used in the grammar are preserved in their
original form. This has the benefit of correct comparison with equals().

For example, instead of providing a new wrapper type for each possible
kind of metadata (such as it had been the case with labels in the form
of LabeledTerm), it suffices to create a new metadata Key and provide a
corresponding method in BnfBuilder.