#2196 Avoid ambiguity in the parser related to yielded expressions
Yield operands can be said to extend regular single operands. This gives
rise to an interesting case: yielded expressions (delimited by
`beginExpr` & `endExpr`) can contain the whole set of yield operands,
not just single operands. Therefore, we need a separate rule for
yielded expressions. However, if it used the same pair of delimiters
(i.e., started with `beginExpr`), an ambiguity would occur: is this a
single operand or a yield operand? This could be resolved with ANTLR's
semantic predicates, but introducing a new pair of delimiters (the
chosen approach) is much simpler.

N.B.: ENDYIELDEXPR was introduced for the sake of consistency, we could
have used BEGINYIELDEXPR ... ENDEXPR as well.