#2196 Splice AND/OR conditions during compilation
To align with EQL's stage 1 structures for AND/OR conditions we need to
transform ANTLR's parse results which are in the form of a binary tree.
Example:
(c1 OR (c2 OR c3)) -> (c1 OR c2 OR c3)

This transformation does not affect the semantics, but it reduces the
depth of resulting trees.