#2197 Enhanced generateCentreQueryCriteriaForTesting to populated and load generated types into GENERATED_CLASSES.
This is the core fix to the problem that was manifesting as a ByteBuddy class loading exception. The problem was related stictly to tests where method CriteriaGenerator.generateCentreQueryCriteriaForTesting was used to build a context for tests, but CriteriaGenerator.generateCentreQueryCriteria was used in companions. And so, because GENERATED_CLASSES was not populated in CriteriaGenerator.generateCentreQueryCriteriaForTestin, a re-generation was attempted, leading to the ByteBuddy error.