#2213 Test erroneous union queries with different numbers of yields
In general, this approach of asserting that an exception is thrown is
far from the best because it ignores potential exceptions that may be
thrown for unexpected reasons. Being more specific and expecting only a
certain type of exceptions doesn't work either since the exception gets
rethrown at some higher level, effectively becoming a cause of another
exception. Should we be making assertions about the stack trace at this
point? I don't think so.