#2249 Provided an example of using AuthorisationValidator + tests.
This example declares 2 properties in TgFuelType with different configurations for AuthorisationValidator. Both properties are assigned their default values in TgFuelTypeDao.new_(), which in turn requires authorisation TgFuelType_CanSaveNew_Token. This authorisation of .new_() demonstrates a nested security scope, allowing creation of an entity instance with default values by a user that has no permission to modify the guarded properties.