#2258 QuerySourceInfoProvider: exclude transient properties of persistent entities
EQL3 metadata was not generated for such properties, hence there was no
need to handle them.

Typeful Metadata does, so they need to be explicitly excluded as they
are irrelevant for EQL and only lead to errors. Specifically, when such
a property is present in QuerySourceInfoProvider but absent from
EqlTable, failure occurs: such a property can end up among query yields,
and then it won't be resolved against EqlTable.