#2258 Simplify PropertyMetadataVisitor
The semantics of method `otherwise` are that of "default value" but that
requires it to always be implemented even when all other visiting
methods are implemented as well -- not nice.

This visitor is not currently used at all, so this change is OK, and
another visitor with "default value" method can be introduced any time.