#2274 Reintroduce `domainTypes` parameter in Guice module constructors
This parameter is necessary for correct dynamic binding of entity
companions. Currently, applications bind companions only for "domain
entity types", as opposed to all entity types. The latter is exposed in
the API of IApplicationDomainProvider, while the former is not. API
change exposing 'domainTypes' should be considered in the future.

This partly reverts 2d1921299015685a72413b8e9b0ad25465840ba2