#2274 Move AuthorisationInterceptor binding to a separate module
Previously it was bound in EntityModule, which hadn't been the best
choice, as there are many platform-level test with their own modules
extend EntityModule but don't make any use of authorisation, with the
exception of one -- AuthorisationProcessTest, which has been explicitly
provided with the relevant binding in this commit.

The new AuthorisationModule is now installed by BasicWebServerModule to
ensure that applications have authorisation enabled, just as before.