#348 Fix a bug in StreamUtils.distinct that caused early stream termination
Spliterator#tryAdvance had not been implemented correctly in our custom
spliterator: it ignored the given action when a non-unique element was
encountered, while returning true (the return value of the original
spliterator). This contradicted the spec, which says:

> If a remaining element exists: performs the given action on it,
> returning true; else returns false.

What had to be done in the case when a non-unique element is encountered
is to keep looking for the next unique one.

IteratorUtils.distinctIterator does exactly this. We then use this
iterator to create a stream of distinct elements, and a spliterator is,
in fact, unnecessary.