#2196 Don't inline rules by default, use a special annotation
Inlining by default goes wrong when we begin inlining more aggressively,
which is required for the SelectSource rule:

SelectFrom = SelectSource as? Join? Where? GroupBy? SelectEnd
SelectSource = select

We would like to inline SelectSource here, but it was not considered
inlineable as it appears in a sequence of other terms. This would
require us to perform aggressive inlining, which could have too
far-reaching effects (e.g., also inlining Join and Where), which could
be addressed with more "NoInline" annotations, but it would get tedious.

With this change we reverse direction w.r.t. inlining - don't do it by
default, but only for rules that are explicitly asking for it. An
upcoming change will add support for aggressive inlining.