#2258 SequencedSet to be returned from Finder.findCommonProperties.
Generally speaking there needs to be some predictable order to properties, returned from various reflection utility methods. Historically, it was assumed that the order should match the order of property declrations in entity classes. There is no guarnatee of such order at the JVM level. However, factually this was the case at least starting with Java 5 and remains in Java 21. TG has the code, such as various tests, that expect such order.