#2275 Corrected finding of key event target.
Previously, keyEventTarget was being found starting from Shadow DOM sub-element ('insertionPointContent') which had tabindex overridden. Now, tabindex is overridden on host element, so we start from it. And actually 'insertionPointContent' does not even exist today. The bad effect was that keyEventTarget was always host element even though sometimes it was meant to be 'tg-menu-item-view' above host.