#1293, #2274 Modified getKeyMemberInfo to return TypeInfo.STRING only for String properties.
The main intent of this change is to support correct conversion of composite keys with a single member to string. Prior to this change composite entities with non-string member, such as Date, would not go through the conversion logic, leading to runtime errors in situations of nested composite enttities. For example, a string representation for composite key in HolidayLocation would fail because Holiday has a single-member composite key of type Date.

One implication of this change is that now any references to a single-membered composite key as property with name 'key' in EQL queries are always converted to a string (VARCHAR). This was not the case before and some queries may rely on the fact that property 'key' in that case would be of the same type at the database leve as a corresponding single member (e.g., datatime2). Such queries need to be adjusted to reference a single key member property explicitly.