#2313 Use QuerySourceInfoProvider to rationalise retrieval model construction
This simplifies construction of retrieval models by removing most of the
conditional statements in include* methods.
Instead, QuerySouceInfoProvider tells us which properties are relevant
to EQL, and thus can be retrieved.

This change also impacts retrieval models for synthetic entities.
Previously, there were hard-coded rules for the fetching of property
`id` which were not clearly understood. Now `id` is fetched only if
QuerySourceInfoProvider says that it belongs to the synthetic entity
(effectively, if its underlying model yields it).

One effect of this change that has to better understood is presence of
crit-only properties in a fetch model. Prior to this change, a fetch
model for a synthetic entity could contain a crit-only property, either
due to having been explicitly added, or added implicitly (based on the
fetch category).
Of course, such a property would also have to be yielded in the
synthetic entity's model.
With this change, crit-only properties are no longer implicitly added to
retrieval models.

In the future, QuerySourceInfoProvider should be refactored into an
extension of IDomainMetadata, which will remove the need for using both
for retrieval model construction.