#2313 Throw an exception if a source query yields into a non-existent property
The main change is in Source1BasedOnQueries.produceQuerySourceInfoForEntityType.

Implementin this required IDomainMetadata to be accessible to
Source1BasedOnQueries. Its `transform` method takes a
TransformationContextFromStage1To2 as an argument, so I considered it
reasonable to enhance that type with a IDomainMetadata component. The
bulk of changes is related to this enhancement.