#348 No need for doubling the length for NVARCHAR.
It turnes out that SQL Server already has byte-pairs semantics for n. The same is true of PostgreSQL in application char, varchar, and text.
At the same time, SQL Server requires MAX to be used for n > 4000. This is why we check for this and issue a warning during DB schema generation.