#2061 Rationalisation of copying of collectional property values
This commit changes method EntityUtils.copyCollectionalValue in 3 ways:

1. The previous version relied on the existence of a public no-arg
constructor in the argument's type. This is true for all collectional
property types used in TG applications. However, their accessors always
return unmodifiable views of the underlying collections, which do not
provide such a constructor. With the recent change to dynamic property
access (fddfce7ed12ab1d4df01e59f346ec24be2b5aae0), this assumption
started causing a runtime error.

The said method is being used only with collectional values obtained via
dynamic property access (method AbstractEntity.get), which previously
read values directly from property fields. Consequently, the expected
constructor was always available.

Now that dynamic property access invokes accessor methods to read
property values, this is no longer true. Thus, a different approach is
required, and is implemented in this commit.

2. Previously, copyCollectionalValue returned a mutable copy; now it
returns an immutable one. This change should be safe, as the
mutability aspect of the copy hadn't been required by the callers.

3. Previously, a returned copy had the same type as the argument.
Now, a copy has a type that is compatible with the collectional type of
the argument. This change should be safe, as the callers hadn't been
relying on this aspect of the previous version.