#2337 More robust copying of collectional property values
This commit changes method EntityUtils.copyCollectionalValue in 3 ways:

1. The previous version relied on the existence of a public no-arg
constructor in the argument's type. This is true for all collectional
property types used in TG applications. However, their accessors always
return unmodifiable views of the underlying collections, which do not
provide such a constructor. With the recent change to dynamic property
access (50bb815d75a68eca3c3ced1d2355e172493cccde), this assumption
started causing a runtime error.

The said method is being used only with collectional values obtained via
dynamic property access, which previously read values directly from
property fields. Consequently, the expected constructor was always
available.

Now that dynamic property access invokes accessor methods, this is no
longer true. Thus, a different approach is required, and is implemented
in this commit.

2. Previously, copyCollectionalValue returned a mutable copy; now it
returns an immutable one. This change should be safe, as the
mutability aspect of the copy hadn't been required by the callers.

3. Previously, a returned copy had the same type as the argument.
Now, a copy has a type that is compatible with the collectional
property's setter. With the standardisation of collectional property
setters, this change should be safe.