#2329 27. Rationalised neon-animation imports.
Only those animations where added that were actually used. And those elements / behaviors that were actually used.

It is not hard to add others to 'build-polymer.js' if required. Even more so, it would be best if every new import in our elements were always supplemented with check to 'build-polymer.js' file. If there are no import in that file -- please add it there. The element can exist (transitive) but adding it to 'build-polymer.js' will ensure nothing is broken if some other dependencies become purged.