#2349 Refactor PropertySpec: use TypeName instead of Type for property type
This change is required for further work - generation of ModProp entity
types. Such types will have a property typed with an audit-entity, which
may not necessarily exist at generation time, thus there will be no Type
object to represent it.

To accomodate for the lack of access to Type, which this change resulted
in, utility method `reflectType` was introduced.