#2349 Dynamically create fetch providers in audit-entity companions
This commit implements it in the base audit-entity companion, removing
the need to embed fetch providers into generated audit-entity companion
types.

Initially, the idea had been to finely craft fetch providers by
including service properties (from AbstractAuditEntity) and all declared
properties except the calculated 'changedProps'. However, domain
metadata cannot provide only declared properties, as this information is
not stored. An attempt was made to enrich metadata with this
information, but it turned out to be a non-trivial task. (Some entities,
such as union entities, have implicit properties, for which metadata is
generated, but which aren't declared in source code.)

As a result, a simple implementation that includes all persistent
properties is provided by this commit. Operationally, it is equivalent
to the initial idea - the same set of properties is included.