#2329 Removed 'package.json' files from dependencies.
All our resolved dependencies, as defined in main `src/main/resources/package.json` file, are listed in 'package-lock.json' file.
This file is regenerated during an update cycle using 'install-polymer.sh|bat'.
Any changes to this file are committed.
The regenerated file shows the actual dependencies and is processed by the Dependabot.

This is why there is no reason to keep 'package.json' files for individual packages.
They may not show the actually loaded sub-dependencies and thus may indicate false-positive vulnerabilities (e.g., 'typescript': '^3.7.5' definition with vulnerability may resolve to '3.7.19' or '3.19.7' which is not vulnerable already).
When such vulnerable sub-dependencies are overridden, these files would not reflect the change and the Dependabot would keep reporting vulnerabilities as if they are present.