#2349 Replace dynamic audit type generation for test entities in favor of manual generation
The dynamic approach that had been used is no longer viable due to the
possibility of existence of multiple audit type versions.
It could be made viable by enhancing it with the ability to detect
whether a new version of audit types needs to be generated, but that is
out of scope of this work.