#2349 Ensure PropertyDescriptor is always created with non-blank property titles
When a blank title used as argument to setKey, validation for `key`
fails due to a violated requiredness constraint (a blank string is
treated as if the value is absent), despite PropertyDescriptor.key
not being annotated with Required, because `key` is always implicitly
required (see AbstractEntity.initProperty).
As a result, PropertyDescriptor.key is null, which leads to NPE in
compareTo().