#2349 Deny introspection of audit types
Persistent audit types are an implementation detail, thus need not be
introspectable.

Synthetic audit-prop types are also an implementation detail, but need
to be introspectable because their data is used in entity centres for
synthetic audit-entities.

It would make sense to refine the concept of introspection so that
metadata operations could be denied while allowing operations on data.