#2381 Always implicitly fetch `desc` with categories ALL, DEFAULT
Prior to this change, categories ALL and DEFAULT were not a superset of
KEY_AND_DESC w.r.t. to property `desc`, as they skipped it when it was
calculated.

Furthermore, in some cases, even when using a sub-fetch model with
category KEY_AND_DESC, `desc` would not be included.
Specifically, when such a sub-fetch model was specified for a key-member
property.
For example, given the following entities:

class Planning extends ... {
    @CompositeKeyMember(1)
    WorkActivity workActivity;
}

@DescTitle
class WorkActivity extends ... {}

Fetching Planning with a fetch model that contains:
* `key` (or simple has category KEY_AND_DESC);
* sub-fetch model for `workActivity` with category KEY_AND_DESC.
resulted in "workActivity.desc" NOT fetched.

This was because fetching `key` implicitly created a sub-fetch model for
the key-member - `workActivity` - with category DEFAULT, which was then
combined with the explicitly specified sub-fetch model
(via `fetch.unionWith`). This union ultimately resulted in category
DEFAULT for `workActivity` sub-fetch model, which wouldn't include
calculated `desc`.