#2381 Corrected a test case that crashed other tests.
The problem manifeted when running all tests with mvn. It was due to the use of a singleton IDomainData instance, which was dynamically polluted with type Circular_EntityWithCompositeKeyMemberUnionEntity, which is not registered as test domain entity (its declaration is insufficient for that purpose). As the result, a TRUNCATE TABLE clause was generated for this entity, without having such table created in the first place. All test cases that get executed after FetchModelTest would fail with a JDBC exception due to table not existing.