#2381 Align generation of DDL and insert/truncate scripts in test
Previously, DDL had been generated using a list of registered entity
types, which is immutable.
Truncate and insert scripts, on the other hand, had been generated using
a list of entity types that have metadata. This list is mutable, as
metadata can be generated ad-hoc.

To avoid situations as described in
54175f8e7ff90d66d12d832c726ef68abaeaaa27, both generation procedures
should use the same source of entity types.