#2381 #2258 Remove method IDomainMetadata.allTypes()
This change strengthens the invariant of the application domain being
statically determined (i.e., new entity types cannot be registered once
the application is running).

Current implementation of domain metadata enables generation of metadata
ad-hoc, which is not consistent w.r.t. the static property of
application domain. Method allTypes() exposed this inconsistency.

Note that this change doesn't address the core issue - ability to
generate metadata ad-hoc for arbitrary entity types. The more thorough
solution would be to restrict this. However, such a change is more
complex and requires deeper analysis. This commit, on the other hand,
is a simpler change, which hides the constistency, lest it be leveraged
in an uncommon use case.

In the future, we may support dynamic registration of entity types.