#2386 Correct stream usage to prevent reuse of a stream that was operated on
Method export, which was renamed to export_ in this commit, was using
the same `propertiesToHyperlinksList` stream in a `for` loop, which
resulted in it being operated on during each iteration, ultimately
causing an error.

This commit modifies the signature of this renamed method to accept a
list of such streams, because the first argument is a list of data
sheets, and each data sheet contain a list of entities. Each
property-to-hyperlink Map is associated with a single entity instance,
hence a list of streams of such maps.