#2388 Limit the use of searchText to persistence only
This commit makes searchText "write-only" with respect to application
and platform code. In Java code it is used only to persist RichText values.

Note that RichTextType (Hibernate type mapper) will retrieve searchText,
and it will be assigned in the RichText constructor. This can be thought
of as an optimisation -- searchText does not have to be computed for
persisted RichText.