#2388 Remove workaround that enabled yielding into RichText properties directly
This workaround enabled the use of a RichText property as a yield alias
directly, just like Money-typed property can be used as a yield alias.

However, there is no practical use for such queries. And the only
benefit of this workaround is the ability to specify "richText" instead
of "richText.searchText". This would make sense only when searchText
needs to be yielded on its own, as a string. There is very little
benefit to gained by this, and it may be confusing to EQL users who
attempt to yield RichText as a whole (recall that the correct way to do
this is to yield formatted and core text separately, and, optionally,
search text).

It would also be good to implement validation to report a corresponding
error.