#2349 Synthetic audit types do not need to yield version
Stricter validation of synthetic models, merged from upstream, was
rejecting models for synthetic audit types due to their yielding
property `version`, and metadata rules say that `version` exists only
for persistent and synthetic-based-on-persistent entity types.

An accompanying change in this commit is in SynAuditEntityUtils:
yieldAll is no longer used.
It is unclear why it had been used in the first place.
Likely with the intention to simplify the code (yieldAll conveniently
returns the right fluent interface).