#185 Force selection to start from zero if queried; for non-empty selected text, that was selected from beginning.
When selecting text from the very end of the editor on a task list or from the middle on a task list using mouse (or keyboard) to the very beginning of text, we still have TransformError or RangeError respectively. Some magic is still present with zero / one selection[0].

This change prevents those errors.