#2349 Refactor synthetic entity metadata by storing models separately
Instead of storing models of synthetic entity types in their
EntityMetadata instances, introduce a separate abstraction --
`ISyntheticModelProvider`.

This change is prompted by the desire to generate models for synthetic
audit types at runtime. In short, such generation would not be possible
otherwise, as it would introduce a circular relationship between the
generator and metadata.