#2349 Runtime generation of models for synthetic audit-entity types
There remains code that does not yet use ISyntheticModelProvider --
ua.com.fielden.platform.eql.meta.EntityTypeInfo.
It could be refactored but that would not be a trivial change.
Therefore, it is put out of scope of this issue.
For compatibility, both approaches of accessing synthetic models are supported right now.
See SynAuditModelInitService for details.

SynAuditModelInitService is a simple way of implementing a service that
needs to be implicitly executed during startup.
If there will arise a need for a proper "startup service" facility, a
more sophisticated approach may be required.

Generator was also modified so that models for synthetic audit-entity
types are no longer generated as source code, only the field is declared.