#2392 Exclude signature files from the uber jar for annotation processors.
Signed dependencies that get included into the uber jar bring it the siganture files, which then breaks the compilation process of TG-based applications with errors reporting an error instantiating annotation processors. The error does not indicate any issues with the uber jar signature.

Also, added minisation of uber jar and updated the compiler and shade plugins.