#2349 Support for disabling of auditing
A few words about the implementation.
There are two obvious ways to go about configuring auditing facilities depending on the auditing mode:

1. Modify an existing implementation by adding assertions about the auditing mode.
2. Bind a "stub" implementation that throws an exception whenever its methods are called.

Approach 1 was chosen. Although the repetetive assertions add a bit of
overhead, it should not have any significant effect. As a plus, no
additional "stub" implementations are needed.

Placing such assertions in constructors is possible, but would require
all users of auditing facilities to switch to using `Provider`s instead
-- to construct them on demand after ensuring that auditing is enabled.

A more interesting problem is the ability to switch auditing modes at
runtime. Currently, this is not supported, as most implementation
classes are singletons, thus `AuditingMode` is injected once and for all.