#2349 Ensure sub-tokens are sorted
This amends e6f9ffb9cd8ec61eb8056981be1eef103bb6b422, which used HashSet
and thus resulted in unpredictable order.

Also, the reason given in the said commit is not accurate.
We can use List as well as Set, since SecurityTokenTreeNodeEntity is not persistent,
so the restriction to Set as the only valid collectional property type does not apply.