#2416 Detach Hibernate proxies when they are no longer needed
The current understanding of the Hibernate loading mechanism is this:
an entity loaded as a Hibernate proxy (via `session.load()`) should not
be detached if it is updated (via `session.update()`).

This change is motivated by an observation of an edge case where a
second loading of an entity could cause a StaleObjectException.
This was primarily caused by concurrent modification, and happended when
the first load was without a lock, but the second one used the UPGRADE
lock.