#2432 Avoid collectional property application in conflicting situation.
In case of a property conflict error on some other editor (than tg-collectional-representor), and for actually concurrently changed collection, we may end-up with a situation where collection List<String> value is getting set into its server side entity. The property is of Collection<SomeEntity> type and such setting is fails.

We should tweak val/origVal pair in modifHolder to avoid that. This commit makes modifHolder with origVal only and EntityResourceUtils value application does not occur.