#2452 Retrieval model construction: explore union members by default
"By default" means "when strategy DEFAULT is used".
This strategy is also used implicitly when an entity-typed property is "explored".

For entity types other than union, key members are explored by default,
as they form an essential part of the structure.
For union types, the essential part is the set of union members.
Therefore, it makes sense to explore the essential parts of union types
as well.