#2459 Improve processing of duplicate key values
Prior to this change, existence of duplicate keys in a source database
would result in invalid entries in `IdCache` -- there would entries with
non-existing IDs, which would ultimately lead to SQL errors regarding
violations of foreign key constraints (due to an attempt to insert a
reference to a non-existing ID).

Now, only the first entry among duplicates is taken, and the rest are ignored.