#2456 Explore entity supertypes for meta-model generation
This change is prompted by an observation that a meta-model for
`AbstractUnionEntity` was not always being generated.
Since this type comes from the platform, it is not present in the
annotation processor's input when compiling a TG application.
A similar problem had been faced at the early stages of meta-model
processor development, when we needed to ensure that a meta-model for
`User` is generated.
Then, we concluded that it would be enough to explore properties of
application entity types.
But the mentioned observation showed that it wasn't enough, since
`AbstractUnionEntity` may not necessarily be used as a property type,
but its meta-model is required, as it is used as a superclass for
meta-models of specific union entity types.

Hence, we should also explore supertypes of domain entity types.