#2349 Replace "versioning information" with "version information"
The term "versioning" denotes the act of managing versions of something
(e.g., software versioning).

In our case, the information is not related to the process of versioning,
but to specific pieces of information about an entity, such as its
version number, who created/updated it and when.
Perhaps, even "version information" is not the best term for this.