#1450 Correct conditions for decrementing `refCount`
The active status of the entity being saved does not matter.
It may as well be inactive.
What matters it the persisted active status.

For example, consider activatable entity A that references activatable entity B:
A.version = 0, A.active = false, A.b = B, B.refCount = 0

Now consider two concurrent operations in this order:
1. A.active = true (activated) => B.refCount += 1.
2. A.active = false (unchanged), A.b = null (dereferenced) => B.refCount -= 1.
   Despite A.active = false, B.refCount must be decremented.