#2515 More consistent empty union tooltips.
It is more consistent with tooltips of other empty entity-typed values.

It shows 'Edit Asset', where 'Asset' is a property title. This is slightly different as for standard entity-typed properties, where we show title of the entity. This is because union entity title may be somewhat technical, like 'Maintenance Capable'.