#2517 Refactored DynamicQueryBuilderSqlTest to become domain-driven.
This commit fixes the problem with intermittent failure of WebApiSecurityTest, which was happening if DynamicQueryBuilderSqlTest would run before. It was binding 'IAuthorisationModel' to 'AuthorisationModelForTests' instead of 'ServerAuthorisationModel'.
This change also corrects tests 'unauthorised_prop_throws_authorisation_exception' and 'unauthorised_prop_initiated_with_mnemonics_throws_authorisation_exception', which were relying on inappropriate for this 'AuthrisationModelForTests'.